requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.20.0
fastapi==0.111.0
python-multipart==0.0.9
reportlab==4.2.0